backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Session==0.8.0
pymongo==4.8.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.0
google-api-python-client==2.103.0
waitress==3.0.2
pytest==8.3.4
selenium==4.19.0
pytz==2025.1
gunicorn==21.2.0
PyJWT==2.8.0